Move Config into Server module
We now move everything server related (like the `Config`) into the server
module. This allows us provide a better outlined API later on. The
binary lives now under `src/bin`, which seems more idiomatic Rust
nowadays.
